extract the core to its own function separate from cli
