+++ deon/cli.py
@@ -1,15 +1,8 @@
 import click
-import os
-from pathlib import Path
 
-import xerox
+from .deon import ExtensionException, FormatException, create
 
-from .parser import Checklist
-from .formats import FORMATS, EXTENSIONS
-
-DEFAULT_CHECKLIST = Path(__file__).parent / 'assets' / 'checklist.yml'
-
-CHECKLIST_FILE = Path(os.environ.get('ETHICS_CHECKLIST', DEFAULT_CHECKLIST))
+from .formats import EXTENSIONS
 
 
 @click.command('deon')
@@ -28,41 +21,20 @@ CHECKLIST_FILE = Path(os.environ.get('ETHICS_CHECKLIST', DEFAULT_CHECKLIST))
                                                                       Default is False , which will append \
                                                                       to existing file.')
 def main(checklist, format, output, clipboard, overwrite):
-    # load checklist
-    cl_path = Path(checklist) if checklist else DEFAULT_CHECKLIST
-    cl = Checklist.read(cl_path)
-
-    output = Path(output) if output else None
-
-    # output extension is given priority if differing format is included
-    if output:
-        # get format by file extension
-        ext = output.suffix.lower()
-        if ext in EXTENSIONS.keys():
-            output_format = EXTENSIONS[ext]
-        else:
-            with click.get_current_context() as ctx:
-                msg = 'Output requires a file name with a supported extension.\n'
+    try:
+        result = create(checklist, format, output, clipboard, overwrite)
+    except ExtensionException:
+        with click.get_current_context() as ctx:
+                msg = 'Output requires a file name with a supported extension.\n\n'
                 raise click.ClickException(msg + ctx.get_help())
-    elif format:
-        if format in FORMATS:
-            output_format = format
-        else:
-            with click.get_current_context() as ctx:
-                msg = "File format is not supported.\n"
+    except FormatException:
+        with click.get_current_context() as ctx:
+                msg = "File format is not supported.\n\n"
                 raise click.ClickException(msg + ctx.get_help())
-    else:
-        output_format = 'markdown'
-
-    template = FORMATS[output_format](cl)
 
     # write output or print to stdout
-    if output:
-        template.write(output, overwrite=overwrite)
-    elif clipboard:
-        xerox.copy(str(template.render()))
-    else:
-        click.echo(template.render())
+    if result:
+        click.echo(result)
 
 
 if __name__ == '__main__':
